+++ NEWS.md
@@ -18,9 +18,13 @@ in output object (`include_levels`) and to remove taxa without levels
 * Application of `subset()` extended to content in slot **relations**.
 * Function `vegtable2kml()` was deprecated to resolve dependencies on GIS
   packages associated to package `plotKML`.
+<<<<<<< HEAD
 * Function `transform()` deprecated.
 * Validity checks allow plots in slot **header** without records in **samples**.
 * Function `summary()` also providing number of plots with records.
+=======
+* Function `transform()` replaced by `cover_trans()`.
+>>>>>>> branch 'master' of git@github.com:kamapu/vegtable.git
 
 ## Bug Fixes